refactor(event): move category hydration helpers to query.go
populateSingleCategories and populateCategories are query helpers that
call ListCategories and ListCategoriesByEventIDs on the storage layer.
They sat under the // Converters section next to the pure FromStorage
mappers, which do no database access.

Move both functions next to their callers Get, GetByUID, and
ListByDateRange in query.go. The move keeps convert.go limited to pure
row-to-domain mapping and keeps every query-shaped helper in one file.
Pure code motion: bodies, receivers, and signatures are unchanged.
